Refact: 부스 상품 서비스와 부스 상품 카테고리 서비스 분리
